requirements.txt
flask==1.1.2
setuptools==45.1.0
flask_restful==0.3.8
marshmallow==2.15.3
Shapely==1.7.1
python-dateutil==2.8.1
requests==2.24.0
flask_wtf==0.14.3
wtforms==2.3.2
flask-markdown==0.3
flask_sqlalchemy==2.4.4
flask_bcrypt==0.7.1
email-validator==1.1.1
flask_login==0.5.0
lxml==4.5.1
xmljson==0.2.1
geojson==2.5.0
